Removed queue swap in remove all data operation.
When the sync manager was removing all sync data, it would cancel
operations in its operation queues, and then create new queues to
continue. This could lead to overlapping operations, and errors in the
console. For this reason, the queues are not replaced. Effectively,
this allows any incomplete ops to finish before the remove data
operation starts.