Changes to lower memory in change sets
Large data objects can cause devices to crash for some change sets.
Changes have been make to try to keep the memory footprint of change
sets low. changedAttributes are now scanned for large NSData objects.
If found, the data is written out to a temporary file, which is mapped
back into memory. Autorelease pools and refreshing of change set
objects should cause the in-memory data to be realloced.